+++ .github/workflows/release.yaml
@@ -1,14 +1,12 @@
-name: Release Charm to Edge and Publish Libraries
+name: Release Charm
 
 on:
-  workflow_dispatch: {}
   push:
     branches:
       - main
+      - track/**
 
 jobs:
   release:
-    uses: canonical/observability/.github/workflows/charm-release.yaml@v0
+    uses: canonical/observability/.github/workflows/charm-release.yaml@v1
     secrets: inherit
-    with:
-      build-for-arm: true
