--- requirements.txt
@@ -1,26 +1,0 @@
-ops ~= 2.5
-pydantic>=2
-
-
-# A collection of helpers and shared code for using Lightkube
-# Code: https://github.com/canonical/lightkube-extensions
-# Deps: charm
-lightkube-extensions @ git+https://github.com/canonical/lightkube-extensions.git@main
-
-# An implementation of the JSON Schema specification
-# Code: https://github.com/python-jsonschema/jsonschema
-# Docs: https://python-jsonschema.readthedocs.io/
-# Deps: traefik_k8s libs, tls_certificates_interface libs
-jsonschema >= 4.22.0
-
-# An implementation of the JSON Schema specification
-# Code: https://github.com/pyca/cryptography/
-# Docs: https://cryptography.io/en/latest/
-# Deps: tls_certificates_interface libs
-cryptography >= 42.0.0
-
-cosl
-
-# PYDEPS
-# lib/charms/tempo_coordinator_k8s/v0/charm_tracing.py
-opentelemetry-exporter-otlp-proto-http==1.21.0
