+++ tox.ini
@@ -4,7 +4,7 @@
 [tox]
 no_package = True
 skip_missing_interpreters = True
-env_list = lint, static, unit, scenario
+env_list = lint, static, unit
 min_version = 4.0.0
 
 [vars]
@@ -12,109 +12,62 @@ src_path = {tox_root}/src
 tests_path = {tox_root}/tests
 ;lib_path = {tox_root}/lib/charms/operator_name_with_underscores
 all_path = {[vars]src_path} {[vars]tests_path}
+uv_flags = --frozen --isolated --extra=dev
 
 [testenv]
+allowlist_externals = uv
 set_env =
     PYTHONPATH = {tox_root}/lib:{[vars]src_path}
     PYTHONBREAKPOINT=pdb.set_trace
     PY_COLORS=1
 pass_env =
     PYTHONPATH
-    CHARM_BUILD_DIR
-    MODEL_SETTINGS
+    CHARM_PATH
+
+
+[testenv:lock]
+description = Update uv.lock with the latest deps
+commands =
+  uv lock --upgrade --no-cache
+
 
 [testenv:fmt]
-description = Apply coding style standards to code
-deps =
-    black
-    ruff
+description = "Format the code"
 commands =
-    black {[vars]all_path}
-    ruff check --fix {[vars]all_path}
+    uv run {[vars]uv_flags} ruff check --fix-only {[vars]all_path}
 
 [testenv:lint]
-description = Check code against coding style standards
-deps =
-    black
-    ruff
-    codespell
+description = Lint the code
 commands =
-    # if this charm owns a lib, uncomment "lib_path" variable
-    # and uncomment the following line
-    # codespell {[vars]lib_path}
-    codespell {tox_root}
-    ruff check {[vars]all_path}
-    black --check --diff {[vars]all_path}
+    uv run {[vars]uv_flags} ruff check {[vars]all_path}
 
 [testenv:unit]
 description = Run unit tests
-deps =
-    pytest
-    pytest-mock
-    coverage[toml]
-    -r {tox_root}/requirements.txt
+allowlist_externals=
+    {[testenv]allowlist_externals}
+    /usr/bin/env
 commands =
-    coverage run \
-      --source={[vars]src_path} \
-      -m pytest -v --tb native --log-cli-level=INFO -s {posargs} {[vars]tests_path}/unit
-    coverage report -m
+    uv run {[vars]uv_flags} coverage run --source={[vars]src_path} -m pytest \
+        {[vars]tests_path}/unit {posargs}
+    uv run {[vars]uv_flags} coverage report
 
 
-[testenv:scenario]
-description = Run scenario tests
-deps =
-    pytest
-    ops-scenario
-    coverage[toml]
-    -r {tox_root}/requirements.txt
+[testenv:static]
+description = Run static checks
+allowlist_externals =
+  {[testenv]allowlist_externals}
+  /usr/bin/env
 commands =
-    coverage run --source={[vars]src_path} \
-                 -m pytest \
-                 --tb native \
-                 -v \
-                 -s \
-                 {posargs} \
-                 {[vars]tests_path}/scenario
-    coverage report
-
-[testenv:static-{charm,lib}]
-description = Run static type checks
-deps =
-    pyright
-    -r {tox_root}/requirements.txt
-commands =
-    charm: pyright {posargs} {[vars]src_path}
-    lib:
+    uv run {[vars]uv_flags} pyright {[vars]all_path}
+    /usr/bin/env sh -c 'for m in $(git diff main --name-only {[vars]lib_path}); do if ! git diff main $m | grep -q "+LIBPATCH\|+LIBAPI"; then echo "You forgot to bump the version on $m!"; exit 1; fi; done'
 
 [testenv:integration]
 description = Run integration tests
-deps =
-    pytest
-    pytest-asyncio
-    pytest-operator
-    juju
-    requests
-    -r {tox_root}/requirements.txt
 commands =
-    pytest -v \
-           -s \
-           --tb native \
-           --log-cli-level=INFO \
-           {posargs} \
-           {[vars]tests_path}/integration
+    uv run {[vars]uv_flags} pytest --exitfirst {[vars]tests_path}/integration {posargs}
+
 
 [testenv:auth-setup]
 description = Run manual auth setup test
-deps =
-    pytest
-    pytest-asyncio
-    pytest-operator
-    juju
-    -r {tox_root}/requirements.txt
 commands =
-    pytest -v \
-           -s \
-           --tb native \
-           --log-cli-level=INFO \
-           {posargs} \
-           {[vars]tests_path}/manual/auth
+    uv run {[vars]uv_flags} pytest --exitfirst {[vars]tests_path}/manual/auth {posargs}
